Generate Obj-C @interface for +c interfaces instead of @protocol
This not only reduces the number of public files exposed to users down to one, it also makes
things safer. It is no longer possible to pass something to a +c interface from Obj-C that is
not compatible and it allows one to directly call static class methods on the generated types
instead of going through the CppProxy type.