Objective-C generated classes copy collection types in constructor
This is a small follow up to #123. The `ObjcGenerator` was already correctly set up to automatically copy `NSString` instances to ensure that if an `NSMutableString` instance was passed instead, an immutable copy was created.
With this change the same is true for `NSSet`, `NSArray`, and `NSDictionary` properties.

Note that thanks to #123 this should not have a performance impact (neither memory or CPU), since at most only one copy will be made.