Use vector to store objects before creating ObjC collection.
Since VLA arrays are stack-based and might fail for very large collections, this is the better approach.

Caveat: Does not work with ObjC @class forward declarations, so I had to change the header hierarchy to import the public header inside the private header. This is quite usual in ObjC so it’s a good change, however might slightly prolong compile time.