Generate default constructors for c++ records
Records can exist as fields within classes. When a record has no
default constructor, it prevents us from using records in classes
that have default constructors. Furthermore, even in cases where
non-default constructors are used, it may not be desireable to
initialize records with specific values at a later point in time.
We could use the brace initializers as a "hack" to get around the
lack of a default constructor, but I would consider using those
types of language constructs to be a poor alternative in this case.