Test const methods, and make const+static rules consistent for Python
Added a test to prove that const methods for C++ work.  While
investigating that, I discovered the validity rules for const and
static, and made them check python the same as Java/ObjC, and had to
remove some unused (and now illegal) code from the Python test suite as a
result.