Fix python cwrapper generation issue
This patch fixes and issue, when multiple source djinni files
are using same kind of compisite object (e.g. list<...>),
which lead to creation of files with the same name more then once,
which resulted in GeneralException. This patch allow reusing of those
files, as the generated context is one and the same.