Fix missing full declarations in IOS files
This patch fixes issue when djinni defined Enum class
was used in some djinni interface, as a key attribute for map or set.
In those cases IOS generated *.mm files were not able to be compiled
without modification of djinni generated content.

In order to fix it the only easy way is to disable forward declarations
for djinni enum types, and use only full header include instead. The hard
and more optimal way would be to extend the djinni with checking if the enum
is actually used by map or set within the interface and include the header file
only then, but that seems to be major refactoring.