Fix Djinni initialization crash in multi shared library case on Android (i.e. djinni_support_lib is a shared library used by other shared libraries). Crash is observed in Android Studio 2.3 with NDK+cmake setup when hitting assert statement: assert(s_singleton) in JniClass (djinni_support.hpp, line 229)
My understanding of the problem is that static_registration class template header is included in two different translation units in the shared libs and so results in bad initialization of the JniClassInitializer map.
Explicit template specialization of static_registration<void *, const JniClassInitializer>::get_map() in the cpp file is a way to force that get_map method to a single translation unit and so enforce its singleton nature.
With the exact same code, init crash was observed on Android 5 but not on Android 6 (tested in both case on a Nexus 5). I have no idea why it's a problem with one Android version and not the other.