Fix generation error with toString and include missing changes.
DESCRIPTION:
There were changes in the ix repo that weren't transfered here. These
printed off a description for toString and a description for cpp enums.
I also added a custom change when generating Java code. For the 'binary'
djinni type, the Java type is 'byte[]'. The toString method tries to
print this off by reference, which throws a Java compilation error.
Instead, in this case print with 'Arrays.toString', passing in the
'byte[]' variable.

TESTING:
Compiled and transfered the .jar file to ix repo. Ran djinni in the ix
repo, and the problems are fixed. That is, the enum descriptions are
being printed, and byte[] printing is implemented correctly.

RELEASE_NOTE:
Fixed java code generation error and include missing changes.

TICKET: