Check for Duplicate Output Paths (#381)
* Check for Duplicate Output Paths

This adds some simple logic to check whether files are set to overwrite each other. This is especially helpful for newcomers who are confused when the default settings cause this issue.

* Simplified Implementation

I realized that the existing codebase actually already does this for files created by the same generator. This change makes it apply to all generators, preventing cross-language overwriting.