Introduce Record.DerivingType.Show
Generating a description of a record is mostly trivial.
In fact, this already happens for Objective-C and Java — only C++ is missing.
Instead of generating this code unconditionally, this commit introduces something equivalent to Haskell’s `Show` typeclass for Djinni records.
It can be written using the regular `deriving (…)` syntax, using the keyword `show`.

Due to how resolver.scala is implemented, this diff is enough to get an error message telling if a nested record isn’t deriving `show` (the message is bad, though).