Only set `-fobjc-arc` for valid languages
`-fobjc-arc` is set as a PUBLIC compile option on djinni_support_lib,
meaning it also gets set for any targets that consume Djinni. However
the Swift compiler doesn't understand this option and will throw an
error if it is present. Any targets that contain Swift code and consume
Djinni, even if indirectly, will thus fail to build.

This change causes the `-fobjc-arc` option to only be set when the
language used for that compilation matches CXX, OBJC, or OBJCXX. CXX is
present because in CMake versions prior to 3.16 or if OBJC/OBJCXX is
not enabled, .m and .mm files are treated as CXX.